corrected integer types in the C <--> FORTRAN interface

+++ src/hash_table.h
@@ -32,7 +32,7 @@ int read_array(const char *file, void *dst, size_t frame_size);
 int write_array(const char *file, void *src, size_t frame_size);
 
 // Fortran callbacks
-extern void delete_dataset(void *ptr, char *dir, int len, int threshold);
+extern void delete_dataset(void *ptr, char *dir, size_t len, int threshold);
 extern int dataset_timeout(void *ptr, int timeout);
 extern void get_channel_range_C(void *ptr, int progress, int *startindex, int *endindex, int *status);
 extern void update_progress_C(void *ptr, int progress);
